chore #16 : 캐러셀 디테일 수정
