+++ Outline/Outline/Source/DesignSystem/Component/PreferenceKey.swift
@@ -1,0 +1,15 @@
+//
+//  PreferenceKey.swift
+//  Outline
+//
+//  Created by Hyunjun Kim on 10/16/23.
+//
+
+import SwiftUI
+
+struct ScrollViewOffsetPreferenceKey: PreferenceKey {
+    static var defaultValue: CGFloat = 0
+    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
+        value = nextValue()
+    }
+}
