+++ Outline/Outline/Source/DesignSystem/Extension/AnimationExtension.swift
@@ -8,6 +8,6 @@
 import SwiftUI
 
 extension Animation {
-    static let openCard = Animation.spring(response: 0.5, dampingFraction: 0.6)
-    static let closeCard = Animation.spring(response: 0.6, dampingFraction: 0.6)
+    static let openCard = Animation.spring(response: 0.4, dampingFraction: 0.6)
+    static let closeCard = Animation.spring(response: 0.4, dampingFraction: 0.5)
 }
